WSS: Add missing cast to aspect ratio calculation
This is C-level stuff, we shouldn't need to do this in C#...

+++ Services/MPExtended.Services.StreamingService/Code/Resolution.cs
@@ -29,7 +29,7 @@ namespace MPExtended.Services.StreamingService.Code
         {
             get
             {
-                return Width / Height;
+                return (double)Width / Height;
             }
         }
